+++ src/components/ProductSwitcher/ProductSwitcher.tsx
@@ -51,7 +51,7 @@ class ProductSwitcher extends React.Component<IProductSwitcherProps, IProductSwi
         super(props);
 
         const { items, startIndex = 0 } = props;
-        const safeStartIndex = startIndex <= items.length - 1 ? startIndex : 0;
+        const safeStartIndex: number = this.getSafeStartIndex(startIndex, items);
 
         this.itemNodes = {};
 
@@ -65,6 +65,25 @@ class ProductSwitcher extends React.Component<IProductSwitcherProps, IProductSwi
 
     componentWillUnmount() {
         window.removeEventListener('resize', this.handleWindowSize, true);
+    }
+
+    componentDidUpdate(prevProps: IProductSwitcherProps) {
+        const { startIndex = 0, items } = this.props;
+
+        if (startIndex !== prevProps.startIndex) {
+            const newIndex: number = this.getSafeStartIndex(startIndex, items);
+            const newValue: string = items[newIndex].value;
+
+            this.setState({
+                currentValue: newValue,
+                currentIndex: newIndex,
+            });
+            this.movePointer(newValue);
+        }
+    }
+
+    getSafeStartIndex(startIndex: number, items: IItem[]): number {
+        return startIndex <= items.length - 1 ? startIndex : 0;
     }
 
     handleWindowSize = () => {